refactor: example
