+++ src/app.mjs
@@ -1,5 +1,20 @@
 'use strict';
 
-import app from './express.mjs';
+import express from './express.mjs';
+import http from 'http';
 
-app.listen(3000);
+let port = 80;
+for (let i = 0; i < process.argv.length; i++) {
+  if (
+    process.argv[i] == '-p' &&
+    process.argv.length > i + 1 &&
+    process.argv[i + 1]
+  ) {
+    port = process.argv[i + 1];
+    i++;
+  }
+}
+
+http.createServer(express).listen(port, () => {
+  console.log(`Server start on port ${port}`);
+});
