+++ src/public/components/dash/button-a.js
@@ -1,0 +1,9 @@
+class CustomButton extends HTMLAnchorElement {
+  constructor() {
+    super();
+  }
+}
+
+customElements.define('button-a', CustomButton);
+
+export default CustomButton;
