+++ src/public/components/dash/master.js
@@ -1,2 +1,3 @@
 import CustomButton from './custom-button.js';
 import InputText from './input-text.js';
+import a from './button-a.js';
